Seperated out the Twitch and TwitchRequest libraries to fix the autoloader issue
