+++ src/Library/Streamlabs.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Stream Labs Library
+ *
+ * Working with streamelements
+ *
+ * @package       API
+ * @author        Marc Towler <marc@marctowler.co.uk>
+ * @copyright     Copyright (c) 2018 Marc Towler
+ * @license       https://github.com/Design-Develop-Realize/api/blob/master/LICENSE.md
+ * @link          https://api.itslit.uk
+ * @since         Version 1.0
+ * @filesource
+ */
+
+namespace API\Library;
+
+
+class Streamlabs
+{
+
+}
